feat(tests): restart outdated docker containers
